Common Nursing Abbreviations
Nursing uses hundreds, if not thousands, of medical acronyms and abbreviations. Test your familiarity with nursing titles, degrees, certifications, and protocols.

Are you unaware of the distinction between a Dx and a Fx? Even seasoned ward nurses are not familiar with all the acronyms and abbreviations used in nursing to refer to various healthcare tests, degrees, certificates, nurse specializations, hospital units, and medical charts and records.

Nonetheless, nurses must be aware of where to find these medical acronyms. The most popular nursing acronyms and healthcare abbreviations are covered in this guide, but not all of them.